[Feat/30/update reservation] 예약 신청 & 거절 api 권한 처리 수정

+++ com.templlo.service.promotion/src/main/java/com/templlo/service/coupon/dto/CouponUseResponseDto.java
@@ -2,7 +2,7 @@ package com.templlo.service.coupon.dto;
 
 import java.math.BigDecimal;
 
-public record CouponUseResponseDto(String status, String message, BigDecimal finalPrice) {
+public record CouponUseResponseDto(String status, String message, Integer finalPrice) {
 	public CouponUseResponseDto(String status, String message) {
 		this(status, message, null); // 기본 생성자: 금액 없는 경우
 	}
